Add files via upload
AdventureGame.cpp is the main structure of the story. The .hpp files were just for me to practice using headers. The Fns.cpp file is for functions and the strings that make up the storyline. I split it up this way so that I could make edits to chunks of the storyline. (I think this will make that easier? That way this is like a template for this type of game.)